Two minor changes in abi-extract.pl:
1. Prettify output by adding a line break after (improved) status
   message.
2. Don't call msgmerge if there is nothing to merge (i.e. avoid an
   error message).


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@30664 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ po/abi-extract.pl
@@ -90,11 +90,15 @@ foreach my $msg (sort keys %strings) {
 
 close OUT;
 
-print "Adding [$ARG1] lines from strings file";
+print "Adding [$ARG1] lines from strings file...\n";
 system("touch tmpmerge.po");
 system("msgcomm --more-than=0 --omit-header --add-location --output=tmp.po tmpmerge.po $file");
-system("msgmerge tmp.po $ARG1.po -o $ARG1.po");
-system("rm tmp.po");
+
+if (-s "tmp.po") {
+    system("msgmerge tmp.po $ARG1.po -o $ARG1.po");
+    system("rm tmp.po");
+}
+
 system("rm tmpmerge.po");
 system("rm $file");
 }
